Create readme-revised.txt
The loop bug is easily remedied for this plug in if a page is first created like in the instructions.  The problem is that this is not clearly stated how important this step is.  There must be another bug that is causing this to happen.  Maybe the use of a conditional within the code for setting page?